MyFunctionApp/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

pandas
matplotlib
numpy
scikit-learn
azure-functions
azure-storage-blob
seaborn
Flask
azure-identity
